backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pymupdf==1.24.5
python-docx==1.1.2
openai==1.30.0
python-dotenv==1.0.0
httpx==0.27.0
python-jose==3.3.0
aiofiles==23.2.1
